Move trackStatisctisc under tableView
